% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_categories_approximate.R
\name{best_categories_approximate}
\alias{best_categories_approximate}
\title{Title}
\usage{
best_categories_approximate(df, category_probabilities,
  encode_cols = NULL)
}
\arguments{
\item{encode_cols}{}
}
\description{
Title
}
