% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_dictionary-data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Recolumnize dictionary example}
\format{A dataframe}
\source{
\href{"https://developer.riotgames.com/apis"}{Riot Games API}
}
\usage{
data(champion_dictionary)
}
\description{
Dictionary corresponding to games_small dataset
}
\examples{
data(champion_dictionary)
}
\keyword{datasets}
