% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-data.R
\docType{data}
\name{games_small}
\alias{games_small}
\title{Recolumnize data example}
\format{A dataframe}
\source{
\href{"https://developer.riotgames.com/apis"}{Riot Games API}
}
\usage{
data(games_small)
}
\description{
Data from 2500 League of Legends games. Collected via official API.
}
\examples{
data(games_small)
}
\keyword{datasets}
