% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recategorize.R
\name{recategorize}
\alias{recategorize}
\title{Title}
\usage{
recategorize(df, category_dictionary, cols_to_encode = NULL,
  ignore_unknown_values = T, handle_duplicate_categories = "first")
}
\arguments{
\item{df}{dataframe that contains the data we want to recategorize}

\item{category_dictionary}{named list with a mapping from values to the categories they belong to}

\item{ignore_unknown_values}{T if unknown values should be ignored, F if they should be one hot encoded}

\item{handle_duplicate_categories}{"first" takes first value in a certain category, "last" takes last value}

\item{enc}{which columns should be altered}
}
\description{
Title
}
\examples{
my_mat <- matrix(c("a","b","c","e","d","o"),ncol = 2, nrow = 3, byrow=TRUE)
dict <- c("v","c","c","c","v","c",
"c","c","v","c","c","c","c","c","v",
"c","c","c","c","c","v","c","c","c","c","c")
names(dict) <- letters
recategorize(my_mat,c(1,2), category_dictionary = dict)
}
