% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encode.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{Title}
\usage{
one_hot_encode(df, encode_cols = NULL, keep = "exists",
  min_occurences = 1)
}
\arguments{
\item{df}{A dataframe that we want to one hot encode}

\item{encode_cols}{The names of the columns that should be encoded, should be categorical}

\item{keep}{"exists" for 1 if value exists, 0 else. "sum" for sum of number of appearances in a row, default exists}

\item{min_occurences}{minimum number of appearances in the data before column is added, default 1.}
}
\value{
dataframe with encode_cols removed and replaced with numeric columns
}
\description{
Title
}
\examples{
one_hot_encode(Titanic,encode_cols = c("Class","Sex","Age"))
}
