% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recategorize.R
\name{recategorize}
\alias{recategorize}
\title{Title}
\usage{
recategorize(df, category_dictionary, cols_to_encode = NULL,
  ignore_unknown_values = T, handle_duplicate_categories = "first")
}
\arguments{
\item{df}{dataframe that contains the data we want to recategorize}

\item{category_dictionary}{named list with a mapping from values to the categories they belong to}

\item{ignore_unknown_values}{T if unknown values should be ignored, F if they should be one hot encoded}

\item{handle_duplicate_categories}{"first" takes first value in a certain category, "last" takes last value}

\item{enc}{which columns should be altered}
}
\description{
Title
}
\examples{
my_mat <- matrix(c("obs1","obs2","obs3","a","a","e","d","o","k"),ncol = 3, nrow = 3)
my_mat <- as.data.frame(my_mat)
colnames(my_mat) <- c("name","1","2")
dict <- rep("consonant",26)
names(dict) <- letters
dict[c("a","e","i","o","u")] <- "vowel"
recategorize(my_mat,cols_to_encode = c(2,3), category_dictionary = dict)
}
