% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recategorize.R
\name{recategorize}
\alias{recategorize}
\title{Recategorize}
\usage{
recategorize(df, category_dictionary, encode_cols = NULL,
  handle_duplicate_categories = "first")
}
\arguments{
\item{df}{dataframe that contains the data we want to recategorize}

\item{category_dictionary}{vector with names consisting of the keys to be looked up and values consisting of the category}

\item{encode_cols}{which columns should be altered}

\item{handle_duplicate_categories}{"first" takes first value in a certain category, "last" takes last value}
}
\value{
dataframe with encode_cols replaced with category columns (NAs are populated if no value in that category is found)
}
\description{
Recategorize
}
\examples{
my_mat <- matrix(c("obs1","obs2","obs3","a","a","k","d","o","e"),ncol = 3, nrow = 3)
my_mat <- as.data.frame(my_mat)
colnames(my_mat) <- c("name","1","2")
dict <- rep("consonant",26)
names(dict) <- letters
dict[c("a","e","i","o","u")] <- "vowel"
my_mat
recategorize(my_mat,encode_cols = c(2,3), category_dictionary = dict)
}
